refactor: test/bus.rs
